test: migrate actions to vitest
