--- cypress/component/DarCollectionTable/actions.spec.jsx
@@ -1,265 +1,0 @@
-import { React } from 'react'
-import Actions from 'src/components/dar_collection_table/Actions'
-import { cloneDeep } from 'src/utils/NodashUtil'
-import { Navigation } from 'src/libs/utils'
-import { Storage } from 'src/libs/storage'
-import EnvironmentUtils from 'src/utils/EnvironmentUtils.js'
-import { BrowserRouter } from 'react-router-dom'
-
-let propCopy
-const collectionId = 1
-const refId1 = '0a4jn-g838d-bsdg8-6s7fs7'
-
-const darColl = {
-  darCollectionId: collectionId,
-  referenceIds: [
-    '4a3fd-g77fd-2f345-4h2g31',
-    '0a4jn-g838d-bsdg8-6s7fs7',
-  ],
-  darCode: 'DAR-9583',
-  name: 'Example DAR 1',
-  submissionDate: '2022-07-26',
-  researcherName: 'John Doe',
-  institutionName: 'Broad Institute',
-  status: 'Draft',
-  hasVoted: false,
-  datasetCount: 4,
-}
-
-const draftDarColl = {
-  darCollectionId: null,
-  referenceIds: [refId1],
-  darCode: 'DRAFT-023',
-  name: null,
-  submissionDate: '2022-07-26',
-  researcherName: null,
-  institutionName: null,
-  status: 'Draft',
-  hasVoted: false,
-  datasetCount: 10,
-}
-
-const user = {
-  userId: 1,
-  roles: [
-    {
-      dacId: 2,
-    },
-    {}, // not all roles are tied to a DAC, this is a stub for those roles
-    {
-      dacId: 3,
-    },
-  ],
-}
-
-const props = {
-  consoleType: 'chair',
-  collection: darColl,
-  showConfirmationModal: () => {},
-}
-
-beforeEach(() => {
-  propCopy = cloneDeep(props)
-  cy.stub(Navigation, 'console').returns({})
-  cy.stub(Storage, 'getCurrentUser').returns(user)
-})
-
-describe('Actions - Container', () => {
-  it('renders the actions container div', () => {
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get('.chair-actions').should('exist')
-  })
-})
-
-describe('Actions - Open Button', () => {
-  it('should render the open button if there is a an Open Action', () => {
-    propCopy.actions = ['Open']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-open-${collectionId}`).should('exist')
-  })
-
-  it('should not render Open Button if there is no valid election for opening/re-opening', () => {
-    propCopy.actions = []
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-open-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Actions - Approve Button', () => {
-  it('should render the approve button if there is an Approve Action', () => {
-    propCopy.actions = ['Approve']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-approve-${collectionId}`).should('exist')
-  })
-
-  it('should not render Approve Button if there is no Approve Action', () => {
-    propCopy.actions = []
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-approve-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Actions - Close Button', () => {
-  it('should render if there is a valid election for canceling (all open elections)', () => {
-    propCopy.actions = ['Cancel', 'Vote']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-cancel-${collectionId}`).should('exist')
-  })
-
-  it('should not render if there is no valid election for canceling (no open elections)', () => {
-    propCopy.actions = []
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-cancel-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Actions - Vote Button', () => {
-  it('should not render if relevant elections are not votable', () => {
-    propCopy.actions = []
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-vote-${collectionId}`).should('not.exist')
-  })
-  it('should render if all relevant elections are votable', () => {
-    propCopy.actions = ['Vote']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-vote-${collectionId}`).should('exist')
-  })
-})
-
-describe('Actions - Update Button', () => {
-  it('should not render if relevant elections are not votable', () => {
-    propCopy.actions = []
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-update-${collectionId}`).should('not.exist')
-  })
-  it('should render if all relevant elections are votable', () => {
-    propCopy.actions = ['Update']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#chair-update-${collectionId}`).should('exist')
-  })
-})
-
-describe('Researcher Actions - Revise Button', () => {
-  it('renders the revise button if the collection is revisable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Revise', 'Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-revise-${collectionId}`).should('exist')
-  })
-  it('does not render if the election is not revisable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-revise-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Researcher Actions - Review Button', () => {
-  it('renders the review button if the collection is reviewable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Revise', 'Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-review-${collectionId}`).should('exist')
-  })
-  it('does not render if the election is not reviewable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Revise']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-review-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Researcher Actions - Resume Button', () => {
-  it('renders the resume button if the collection is resumable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Resume', 'Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-resume-${collectionId}`).should('exist')
-  })
-  it('does not render if the election is not resumable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-resume-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Researcher Actions - Delete Button', () => {
-  it('renders the delete button if the collection is deletable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Delete', 'Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-delete-${collectionId}`).should('exist')
-  })
-  it('does not render if the election is not deletable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-delete-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Researcher Actions - Draft', () => {
-  it('uses the referenceId in id if draft', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.collection = draftDarColl
-    propCopy.actions = ['Revise', 'Resume', 'Review', 'Cancel', 'Delete']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-delete-${collectionId}`).should('not.exist')
-    cy.get(`#researcher-resume-${refId1}`).should('exist')
-    cy.get(`#researcher-review-${refId1}`).should('exist')
-    cy.get(`#researcher-cancel-${refId1}`).should('exist')
-    cy.get(`#researcher-delete-${refId1}`).should('exist')
-    cy.get(`#researcher-revise-${refId1}`).should('exist')
-  })
-})
-
-describe('Researcher Actions - Update Button', () => {
-  it('renders the update button if the collection is updatable', () => {
-    cy.stub(Storage, 'getEnv').returns(EnvironmentUtils.envGroups.NON_PROD[0])
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Resume', 'Create_Progress_Report']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-create-progress-report-${collectionId}`).should('exist')
-  })
-  it('does not render if the collection is not updatable', () => {
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-create-progress-report-${collectionId}`).should('not.exist')
-  })
-})
-
-describe('Researcher Actions - Review Closeout Button', () => {
-  it('renders the review closeout button if the collection has Review_Progress_Report action in non-prod environment', () => {
-    cy.stub(Storage, 'getEnv').returns(EnvironmentUtils.envGroups.NON_PROD[0])
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review_Progress_Report']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-review-closeout-${collectionId}`).should('exist')
-  })
-
-  it('does render in production environment even with Review_Progress_Report action', () => {
-    cy.stub(Storage, 'getEnv').returns('prod')
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review_Progress_Report']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-review-closeout-${collectionId}`).should('exist')
-  })
-
-  it('does not render if Review_Progress_Report action is not present', () => {
-    cy.stub(Storage, 'getEnv').returns(EnvironmentUtils.envGroups.NON_PROD[0])
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-review-closeout-${collectionId}`).should('not.exist')
-  })
-
-  it('renders with correct label "Review Closeout"', () => {
-    cy.stub(Storage, 'getEnv').returns(EnvironmentUtils.envGroups.NON_PROD[0])
-    propCopy.consoleType = 'researcher'
-    propCopy.actions = ['Review_Progress_Report']
-    cy.mount(<BrowserRouter><Actions {...propCopy} /></BrowserRouter>)
-    cy.get(`#researcher-review-closeout-${collectionId}`).should('contain.text', 'Review Closeout')
-  })
-})
